chore: wxt 설치 및 entrypoints 정리
