+++ .gitignore
@@ -19,10 +19,16 @@ node_modules
 dist
 dist-ssr
 *.local
+.output
+stats.html
+stats-*.json
+.wxt
+web-ext.config.ts
 
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
+!.vscode/settings.json
 .idea
 .DS_Store
 *.suo
