+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## `v0.3.6` (2025-09-02)
+
+- 수업 목록(`courseIds`)에 대한 만료 검사가 없어서 추가
+
 ## `v0.3.5` (2025-09-01)
 
 - 학기 바뀌고 이전 데이터 남아있는 문제 수정
