+++ entrypoints/sidepanel/App.tsx
