+++ entrypoints/sidepanel/Router.tsx
