+++ entrypoints/sidepanel/components/AssignmentList.tsx
