+++ entrypoints/sidepanel/components/Header.tsx
