+++ entrypoints/sidepanel/components/common/CourseItem.tsx
