+++ entrypoints/sidepanel/eslint.config.js
