+++ entrypoints/sidepanel/hooks/useCourseAssignment.ts
