+++ entrypoints/sidepanel/hooks/useCourseVod.ts
