+++ entrypoints/sidepanel/index.css
