+++ entrypoints/sidepanel/index.html
