+++ entrypoints/sidepanel/layouts/RootLayout.tsx
