+++ entrypoints/sidepanel/main.tsx
