+++ entrypoints/sidepanel/pages/Assignment.tsx
