+++ entrypoints/sidepanel/pages/HomePage.tsx
