+++ entrypoints/sidepanel/pages/Settings.tsx
