+++ entrypoints/sidepanel/types/courseAssignmentData.ts
