+++ entrypoints/sidepanel/types/courseVodData.ts
