+++ entrypoints/sidepanel/vite-env.d.ts
