--- manifest.json
@@ -1,27 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/chrome-manifest",
-  "manifest_version": 3,
-  "name": "이코 (이클래스 도와주는 코끼리)",
-  "description": "동국대학교 이클래스 보조 확장 프로그램입니다.",
-  "version": "0.3.6",
-  "side_panel": {
-    "default_path": "index.html"
-  },
-  "background": {
-    "service_worker": "service-worker/index.js",
-    "type": "module"
-  },
-  "permissions": ["scripting", "storage", "tabs", "sidePanel"],
-  "host_permissions": ["https://eclass.dongguk.edu/*"],
-  "content_scripts": [
-    {
-      "matches": ["https://eclass.dongguk.edu/*"],
-      "js": ["content-scripts/sidePanelButton.js"]
-    }
-  ],
-  "icons": {
-    "16": "assets/icons/icon16.png",
-    "48": "assets/icons/icon48.png",
-    "128": "assets/icons/icon128.png"
-  }
-}
